replace ruleset construction in connection manager

+++ quisp/rules/Rule.cc
@@ -9,7 +9,7 @@ void Rule::setCondition(std::unique_ptr<Condition> cond) { condition = std::move
 
 void Rule::setAction(std::unique_ptr<Action> act) { action = std::move(act); }
 
-void Rule::setNextRule(unsigned long next_rule_id) {
+void Rule::setNextRule(int next_rule_id) {
   if (to != -1) {
     throw omnetpp::cRuntimeError("next_rule_id has already been set");
   } else {
